fix: log dir config, update-check 403, telnet crash, taskbar icon
- Settings -> General: configurable log file directory (Select/Reset),
  applied at startup after the profile loads
- HttpHelper: send User-Agent so GitHub API does not 403 the update check
- Telnet: fix P/Invoke out_len int/nint mismatch (Rust usize is 8 bytes)
  that corrupted the poll buffer and closed sessions immediately
- Taskbar icon always uses the app LOGO (session icons are too small and
  looked blurry), session icons remain on tab headers
- ssh-rust: add fake-telnet-server integration test + out_len 64-bit guard
- AppIcons: cached window icon loaded from compiled LOGO.ico/LOGO_D.ico

release: bump to 1.0.6

+++ Ui/Service/RustSsh/SshRustNative.cs
@@ -84,7 +84,7 @@ namespace _1RM.Service.RustSsh
             long handle,
             [Out] byte[] buf,
             int cap,
-            out int outLen);
+            out nint outLen);
 
         /// <summary>
         /// Notify the remote PTY of a resize.
